Python-Project-Structure/requirements.txt
alembic==1.5.6
aniso8601==9.0.1
blinker==1.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Faker==6.6.0
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
text-unidecode==1.3
Werkzeug==1.0.1
WTForms==2.3.3
